+++ lib/stages/BaseWriter.hpp
@@ -134,7 +134,7 @@ protected:
         std::map<uint32_t, uint32_t> freq_id_map;
 
         // Data size
-        int32_t frame_size = -1;
+        int64_t frame_size = -1;
 
         /// Last update
         double last_update;
